Regenerate image.c every make call
With every 'make' call, the image.c will be regenerated from an image.png
which should be placed in the root directory. This is useful to avoid one
command call.

Instructions and sample file modified accordingly, to test, simply run make
and see the image.c file generated and integrated into Joystick.hex.